FEAT: added reviewerUI
